+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -321,7 +321,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSS3K2VJFMD1JMX5EE8SCZM1"
+        "01KSSB070V0YPWTKJG6TK4RD2Y"
     )
     r"""The unique identifier of the data source"""