+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSS3K2VGZ92BN96RRMSJKD9K"
+        "01KSSB070SQQ3ZNRARZFHWHXPJ"
     )
     r"""The unique identifier of the data source"""