+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSS3K2VH3QVMDC6W7B0ZSYCW"
+        "01KSSB070T3YH1QNPF50CM4GKG"
     )
     r"""The unique identifier of the data source"""