+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -154,7 +154,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-05-29T05:33:39.324Z")
+    updated: Optional[datetime] = parse_datetime("2026-05-29T07:43:08.807Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")