+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSS3K2VMCNVDG5DQ3JM9A08C"
+        "01KSSB070YK7YE81MWNAAQMZ2H"
     )
     r"""The unique identifier of the data source"""